VS8.0 has two implementations of div(). Let's explicitly use the int-aware one.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@23223 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/extlib/ib_udf.cpp
@@ -136,7 +136,7 @@ double EXPORT IB_UDF_cot( double *a)
 double EXPORT IB_UDF_div( long *a, long *b)
 {
 	if (*b != 0) {
-		div_t div_result = div(*a, *b);
+		div_t div_result = div((int) *a, (int) *b);
 		return (div_result.quot);
 	}
 	else
@@ -248,7 +248,7 @@ char *EXPORT IB_UDF_ltrim( const char *s)
 double EXPORT IB_UDF_mod( long *a, long *b)
 {
 	if (*b != 0) {
-		div_t div_result = div(*a, *b);
+		div_t div_result = div((int) *a, (int) *b);
 		return (div_result.rem);
 	}
 	else
